refactor(discovery): split discovery.go into per-concern files
Apply the same per-concern file-split pattern as #253 (controllers)
and #255 (orchestrator) to the discovery package. The 576 LOC
discovery.go mixed five concerns plus utility helpers.

New layout:

  pkg/discovery/
    discovery.go    (Result, Config, Run, mergeParents, discoverer struct, loadManifests, loadAt)
    bootstrap.go    (seedBootstrapSource + aliasBootstrapSources — synthetic GitRepository wiring)
    rsexpand.go     (renderResourceSet + resolveInputProvider — ResourceSet rendering during load)
    orphans.go      (promoteOrphans — file-indexed existence resolution)
    paths.go        (ResolveScanPath + FindRepoRoot + stripDotSlash + pathUnderRoot)
    remotes.go      (unchanged — bootstrap URL-match)

Net: discovery.go drops 576 → 260 LOC, focused on the discovery Run
flow. No behavior change; identical imports re-distributed.

Verified:
- go test -count=1 ./...      — full suite green
- golangci-lint run ./...     — 0 issues
- tholinka/home-ops           — 118/0
- JJGadgets/Biohazard         — 199/0
- m00nwtchr/homelab-cluster   — 287/2 (pre-existing 404)
- buroa/k8s-gitops            — 73/0
- Zariel/home-ops             — 93/0
