+++ pkg/discovery/discovery.go
@@ -16,18 +16,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/fs"
 	"log/slog"
 	"os"
 	"path/filepath"
 
-	sourcev1 "github.com/fluxcd/source-controller/api/v1"
-
-	fluxopv1 "github.com/controlplaneio-fluxcd/flux-operator/api/v1"
-
 	"github.com/home-operations/flate/pkg/loader"
 	"github.com/home-operations/flate/pkg/manifest"
-	"github.com/home-operations/flate/pkg/resourceset"
 	"github.com/home-operations/flate/pkg/store"
 )
 
@@ -150,30 +144,6 @@ type discoverer struct {
 	sourceFiles map[manifest.NamedResource]string
 }
 
-// seedBootstrapSource publishes a synthetic GitRepository pointing at
-// the working tree's repo root — the anchor for spec.path resolution
-// when a Kustomization carries no explicit sourceRef.
-func (d *discoverer) seedBootstrapSource() (string, error) {
-	abs, err := ResolveScanPath(d.cfg.Path)
-	if err != nil {
-		return "", err
-	}
-	root := FindRepoRoot(abs)
-
-	repo := &manifest.GitRepository{
-		Name: manifest.BootstrapSourceID.Name, Namespace: manifest.BootstrapSourceID.Namespace,
-		GitRepositorySpec: sourcev1.GitRepositorySpec{URL: "file://" + root},
-	}
-	id := repo.Named()
-	d.cfg.Store.AddObject(repo)
-	d.cfg.Store.SetArtifact(id, &store.SourceArtifact{
-		Kind: manifest.KindGitRepository,
-		URL:  repo.URL, LocalPath: root,
-	})
-	d.cfg.Store.UpdateStatus(id, store.StatusReady, "bootstrap")
-	return root, nil
-}
-
 // loadManifests scans cfg.Path, then iteratively follows each loaded
 // Flux Kustomization's spec.path until a fixed point is reached.
 // Interleaved with KS expansion is ResourceSet rendering: a parent
@@ -288,289 +258,3 @@ func (d *discoverer) loadAt(ctx context.Context, l *loader.Loader, dir string, s
 	return nil
 }
 
-// renderResourceSet evaluates rs.Spec across its inputs and AddObjects
-// every resulting recognized Flux resource into the store. The rendered
-// children are attributed to the ResourceSet's own source file so the
-// change filter treats them as siblings of the ResourceSet definition
-// (a ResourceSet change reruns its children's reconciles). Returns
-// the count of new objects added so the caller can detect a fixed
-// point in the expansion loop.
-func (d *discoverer) renderResourceSet(rs *manifest.ResourceSet) (int, error) {
-	docs, err := resourceset.Render(rs, d.resolveInputProvider)
-	if err != nil {
-		return 0, err
-	}
-	srcFile := d.sourceFiles[rs.Named()]
-	opts := manifest.ParseDocOptions{WipeSecrets: d.cfg.WipeSecrets}
-	added := 0
-	for _, doc := range docs {
-		obj, err := manifest.ParseDoc(doc, opts)
-		if err != nil {
-			slog.Debug("resourceset: skipped doc", "rs", rs.NamespacedName(), "err", err)
-			continue
-		}
-		if _, ok := obj.(*manifest.RawObject); ok {
-			// Generic / unrecognized kinds: not something flate
-			// reconciles further. Skipped here; the orchestrator's
-			// post-Run RS expansion pass picks them up and attributes
-			// them to the owning KS for `flate build` visibility.
-			// That late pass sees RSIPs emitted from KS reconcile
-			// (kustomize-substituted dragonfly-${APP} style) which
-			// this discovery pass would miss.
-			continue
-		}
-		id := obj.Named()
-		if d.cfg.Store.GetObject(id) != nil {
-			continue
-		}
-		d.cfg.Store.AddObject(obj)
-		if srcFile != "" {
-			d.sourceFiles[id] = srcFile
-		}
-		added++
-	}
-	return added, nil
-}
-
-// resolveInputProvider satisfies resourceset.ProviderResolver against
-// the discoverer's object store. Name lookups hit a single id;
-// selector matches walk the store's RSIPs in the requested namespace
-// and filter by metadata.labels.
-func (d *discoverer) resolveInputProvider(ref fluxopv1.InputProviderReference, namespace string) ([]*manifest.ResourceSetInputProvider, error) {
-	if ref.Name != "" {
-		id := manifest.NamedResource{
-			Kind:      manifest.KindResourceSetInputProvider,
-			Namespace: namespace,
-			Name:      ref.Name,
-		}
-		obj, _ := d.cfg.Store.GetObject(id).(*manifest.ResourceSetInputProvider)
-		if obj == nil {
-			return nil, nil
-		}
-		return []*manifest.ResourceSetInputProvider{obj}, nil
-	}
-	if ref.Selector == nil {
-		return nil, nil
-	}
-	var out []*manifest.ResourceSetInputProvider
-	for _, obj := range d.cfg.Store.ListObjects(manifest.KindResourceSetInputProvider) {
-		p, ok := obj.(*manifest.ResourceSetInputProvider)
-		if !ok || p.Namespace != namespace {
-			continue
-		}
-		match, err := resourceset.MatchSelector(ref.Selector, p.Labels)
-		if err != nil {
-			return nil, err
-		}
-		if match {
-			out = append(out, p)
-		}
-	}
-	return out, nil
-}
-
-// promoteOrphans materializes Existence entries that no Kustomization
-// will ever render. An "orphan" here is a file-indexed object whose
-// source file is not under any loaded KS's spec.path — render-driven
-// discovery would leave it stranded otherwise. Common shapes:
-//
-//   - A loose HelmRelease/source at repo root that `flate build`
-//     should still render even with no enclosing KS.
-//   - flux-system bootstrap CRs that exist beside the bootstrap KS
-//     itself but aren't inside its spec.path tree.
-//
-// Entries whose path IS under a KS's spec.path stay in Existence —
-// they'll be emitted by that KS's render via emitRenderedChildren
-// (or resolved on-demand by depwait's lazy-promotion fallback when a
-// substituteFrom edge fires before the producing KS reconciles).
-//
-// Shares the parent-path-prefix predicate with the orchestrator's
-// detectOrphans via loader.KSPathPrefixes — the two run in
-// complementary phases (pre-reconcile materialization here,
-// post-reconcile failure demotion there) and must agree on the
-// same "under a KS path" predicate so an object can't be classified
-// orphan by one and parented by the other.
-func (d *discoverer) promoteOrphans() {
-	if d.loader.Existence == nil {
-		return
-	}
-	prefixes := loader.KSPathPrefixes(d.cfg.Store)
-	for id := range d.loader.Existence.All() {
-		if d.cfg.Store.GetObject(id) != nil {
-			continue
-		}
-		file, hasFile := d.sourceFiles[id]
-		if hasFile {
-			if _, covered := loader.LongestParent(prefixes, file, id); covered {
-				continue
-			}
-		}
-		if !d.loader.Existence.Promote(d.cfg.Store, id, d.cfg.WipeSecrets) {
-			slog.Debug("discovery: orphan promotion failed", "id", id.String())
-		}
-	}
-}
-
-// aliasBootstrapSources seeds a working-tree SourceArtifact for every
-// GitRepository referenced by a loaded Kustomization whose definition
-// isn't in the repo itself. Targets `flux bootstrap` / flux-operator
-// FluxInstance patterns: the cluster's root GitRepository is created
-// out-of-band (the source that delivers the rest of the manifests
-// cannot, by construction, be one of the manifests it delivers), so
-// no static manifest exists in the tree to discover. Without aliasing,
-// every Kustomization referencing it via `sourceRef` fails depwait
-// with "dependency not found" (issue #199).
-//
-// All namespaces are aliased, not just `flux-system` — the convention
-// of running Flux in a non-default namespace (e.g. `gitops-system`)
-// is widespread, and the bootstrap-source-points-at-the-local-tree
-// property is identical regardless of where the user happens to deploy
-// Flux. A typo'd sourceRef name will silently render against the
-// working tree instead of failing fast — same trade-off the
-// `flux-system` path already accepted.
-func (d *discoverer) aliasBootstrapSources(repoRoot string) {
-	known := make(map[manifest.NamedResource]struct{})
-	for _, obj := range d.cfg.Store.ListObjects(manifest.KindGitRepository) {
-		known[obj.Named()] = struct{}{}
-	}
-	seen := make(map[manifest.NamedResource]struct{})
-	var aliased []manifest.NamedResource
-	for _, obj := range d.cfg.Store.ListObjects(manifest.KindKustomization) {
-		ks, ok := obj.(*manifest.Kustomization)
-		if !ok || ks.SourceKind != manifest.KindGitRepository {
-			continue
-		}
-		id := manifest.NamedResource{Kind: manifest.KindGitRepository, Namespace: ks.SourceNamespace, Name: ks.SourceName}
-		if _, ok := known[id]; ok {
-			continue
-		}
-		if _, dup := seen[id]; dup {
-			continue
-		}
-		seen[id] = struct{}{}
-		alias := &manifest.GitRepository{
-			Name: id.Name, Namespace: id.Namespace,
-			GitRepositorySpec: sourcev1.GitRepositorySpec{URL: "file://" + repoRoot},
-		}
-		d.cfg.Store.AddObject(alias)
-		d.cfg.Store.SetArtifact(id, &store.SourceArtifact{
-			Kind: manifest.KindGitRepository,
-			URL:  alias.URL, LocalPath: repoRoot,
-		})
-		d.cfg.Store.UpdateStatus(id, store.StatusReady, "bootstrap alias")
-		slog.Debug("discovery: aliased bootstrap GitRepository",
-			"id", id.String(), "localPath", repoRoot)
-		aliased = append(aliased, id)
-	}
-
-	// Second pass: a GitRepository CR may be file-loaded AND point
-	// at the same remote that flate is being run against (the
-	// Zariel/home-ops pattern — the in-tree GitRepository URL is
-	// the cluster's own deploy-key SSH URL). Real Flux fetches that
-	// URL with a SOPS-decrypted key; flate runs offline, so the
-	// fetch either round-trips slowly over the network or fails on
-	// the wiped-to-placeholder key. URL-match those against the
-	// working tree's git remotes and override the artifact to use
-	// the working tree directly. Single explicit warn when this
-	// fires so misattributed in-tree GitRepositories (a coincidental
-	// URL match on a real shared-infra repo) are spottable in logs.
-	remotes := readWorkingTreeRemotes(repoRoot)
-	debugLogRemotes(remotes)
-	if len(remotes) > 0 {
-		for _, obj := range d.cfg.Store.ListObjects(manifest.KindGitRepository) {
-			repo, ok := obj.(*manifest.GitRepository)
-			if !ok {
-				continue
-			}
-			id := repo.Named()
-			if _, alreadyAliased := seen[id]; alreadyAliased {
-				continue
-			}
-			normalized := normalizeGitURL(repo.URL)
-			if normalized == "" {
-				continue
-			}
-			if _, match := remotes[normalized]; !match {
-				continue
-			}
-			seen[id] = struct{}{}
-			d.cfg.Store.SetArtifact(id, &store.SourceArtifact{
-				Kind: manifest.KindGitRepository,
-				URL:  "file://" + repoRoot, LocalPath: repoRoot,
-			})
-			d.cfg.Store.UpdateStatus(id, store.StatusReady, "bootstrap alias (URL matches working tree)")
-			slog.Info("discovery: aliased in-tree GitRepository to working tree (URL matches working-tree remote)",
-				"id", id.String(), "url", repo.URL, "normalizedKey", normalized, "localPath", repoRoot)
-			aliased = append(aliased, id)
-		}
-	}
-	// Multiple unresolved GitRepositories is the cross-repo footgun:
-	// each gets aliased to the SAME working tree, so a real upstream
-	// shared-infra GitRepository would render against the wrong files
-	// without any user-visible signal. Warn so an operator can spot
-	// the divergence; the single-source case stays Debug because
-	// that's the intended flux-bootstrap shape.
-	if len(aliased) > 1 {
-		names := make([]string, len(aliased))
-		for i, id := range aliased {
-			names[i] = id.String()
-		}
-		slog.Warn("discovery: aliased multiple GitRepositories to the working tree; cross-repo refs render against the wrong tree",
-			"count", len(aliased), "ids", names, "localPath", repoRoot)
-	}
-}
-
-// ResolveScanPath normalizes a user-supplied --path / --path-orig:
-// absolute, with symlinks resolved. Without symlink resolution
-// filepath.WalkDir doesn't follow root-level symlinks, producing an
-// empty manifest set without any error indication — a footgun.
-func ResolveScanPath(p string) (string, error) {
-	abs, err := filepath.Abs(p)
-	if err != nil {
-		return "", err
-	}
-	resolved, err := filepath.EvalSymlinks(abs)
-	if err != nil {
-		if errors.Is(err, fs.ErrNotExist) {
-			return abs, nil
-		}
-		return "", fmt.Errorf("resolve --path %q: %w", p, err)
-	}
-	return resolved, nil
-}
-
-// FindRepoRoot walks upward from p looking for a .git directory; falls
-// back to p itself when there isn't one.
-func FindRepoRoot(p string) string {
-	for cur := p; ; {
-		if _, err := os.Stat(filepath.Join(cur, ".git")); err == nil {
-			return cur
-		}
-		parent := filepath.Dir(cur)
-		if parent == cur {
-			return p
-		}
-		cur = parent
-	}
-}
-
-func stripDotSlash(p string) string {
-	for len(p) > 0 && (p[0] == '.' || p[0] == '/') {
-		if p[0] == '.' && (len(p) == 1 || p[1] == '/') {
-			p = p[1:]
-			continue
-		}
-		if p[0] == '/' {
-			p = p[1:]
-			continue
-		}
-		break
-	}
-	return p
-}
-
-func pathUnderRoot(target, root string) bool {
-	t := filepath.Clean(target) + string(filepath.Separator)
-	r := filepath.Clean(root) + string(filepath.Separator)
-	return len(t) >= len(r) && t[:len(r)] == r
-}
